+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.3.0] - 2018-12-11
+
 ### Added
 
 - [github.com/goph/emperror](https://github.com/goph/emperror) compatible error handler
@@ -36,5 +39,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial release
 
 
-[Unreleased]: https://github.com/goph/logur/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/goph/logur/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/goph/logur/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/goph/logur/compare/0.1.0...v0.2.0